Update README.md
Established in the year 2012, Suresh Gyan Vihar University set out its goal to develop an educated society for the growth of the nation. We are proud to be the center of Distance and Online Education (CDOE) for decades and to groom leaders, not followers who are not just competent but also morally upright. We are the first private University in Rajasthan to be accredited "A" Grade by NAAC.
Due to lesser funds and monetary support, most of the influencers become dropouts, and to fulfill their dreams Suresh Gyan Vihar University (SGVU) offers distance education programs – BA, B.com, BBA, MA, M.com, MBA, and BA (JOURNALISM)  that are designed for working professional and helps them to have a Degree and a Salary at the same time. Following the benefits of Distance Education and Online Learning, SGVU is conducting, controlling, and monitoring online classes and programs by professional staff members and officers to maintain the objective and standard that upgrade Distance Education to the next level. https://distance.sgvu.org/